requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langchain-core
faiss-cpu
python-docx
pdfplumber
python-dotenv
scikit-learn
google-generative-ai
tqdm
pypdf
pytest
reportlab
python-multipart
